My supervisor told me to convert the CLI app functions into endpoints for testing, I updated transaction.py to contain key functions, and api.py for the endpoints

+++ balance.py
@@ -4,7 +4,7 @@ class BalanceTracker:
     def __init__(self):
         try:
             with open("balance.txt", "r") as balance_data:
-                self.balance = int(balance_data.read())
+                self.balance = float(balance_data.read())
         except FileNotFoundError:
             self.balance = 0
